🩹 PromotionTag 전역적 사용을 위해 따로 모델을 분리
